HTML5 Markup and CSS fixed as a quick solution.
This is  a quick solution to fixed some of the markup and the style
sheet.
Now! there is no error on the web page (index.html) and is valid as a
HTML5 document, however this is a quick solution, in fact the web page
need to rebuild from scratch and apply responsive design, which will be
publish soon.